pyproject.toml
[project]
name = "columbia-ai-agents-intro-lecture"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "litellm>=1.80.0",
    "python-dotenv>=1.2.1",
]
